refactor: /features/main에 index.ts 추가 및 파일 구조 변경
