+++ src/features/main/index.ts
@@ -1,0 +1,8 @@
+export * from './card-announcement'
+export * from './card-event'
+export * from './section-announcement'
+export * from './section-banner'
+export * from './section-bottom'
+export * from './section-event'
+export * from './section-like-course'
+export * from './section-new-course'
